it appears to be working

+++ backend/app/types/users/users.resolvers.js
@@ -1,6 +1,13 @@
+const { AuthenticationError } = require('apollo-server');
+const bcrypt = require('bcrypt');
+const jsonwebtoken = require('jsonwebtoken');
 const { User, Order, Pizza, Topping, Size } = require('../../models');
+const { JWT_SECRET } = require('../../config');
 
 function users(_, args, ctx) {
+  if (!ctx.admin) {
+    throw new AuthenticationError('Only admins can query all users');
+  }
   return User.findAll({
     include: [Order],
     order: [[Order, 'createdAt', 'DESC']]
@@ -8,6 +15,9 @@ function users(_, args, ctx) {
 }
 
 async function user(_, args, ctx) {
+  if (!ctx.user || ctx.user.id != args.id) {
+    throw new AuthenticationError('You are not allowed to query that user');
+  }
   const user = await User.findOne({
     where: { id: args.id },
     include: [
@@ -26,21 +36,93 @@ async function user(_, args, ctx) {
   return user;
 }
 
-function newUser(_, args, ctx) {
-  return User.create({ ...args.input });
+async function myProfile(_, args, ctx) {
+  if (!ctx.user) {
+    throw new AuthenticationError('You must login first.');
+  }
+  const user = await User.findOne({
+    where: { id: ctx.user.id },
+    include: [
+      {
+        model: Order,
+        include: [
+          {
+            model: Pizza,
+            include: [Topping, Size]
+          }
+        ]
+      }
+    ],
+    order: [[Order, 'createdAt', 'DESC']]
+  });
+  return user;
 }
 
 function removeUser(_, args, ctx) {
+  if (!ctx.user || ctx.user.id !== args.id) {
+    throw new AuthenticationError('You are not allowed to remove that user!');
+  }
   return User.destroy({ where: { id: args.id } });
+}
+
+// Handle user signup
+async function signup(_, args, ctx) {
+  const { input } = args;
+  const { firstName, lastName, email, password } = input;
+  const user = await User.create({
+    firstName,
+    lastName,
+    email,
+    password: await bcrypt.hash(password, 10)
+  });
+
+  // return json web token
+  return {
+    token: jsonwebtoken.sign(
+      { user: { id: user.id, email: user.email } },
+      JWT_SECRET,
+      {
+        expiresIn: '1y'
+      }
+    )
+  };
+}
+
+async function login(_, args, ctx) {
+  const { email, password } = args;
+  const user = await User.findOne({ where: { email } });
+
+  if (!user) {
+    throw new Error('No user with that email');
+  }
+
+  const valid = await bcrypt.compare(password, user.password);
+
+  if (!valid) {
+    throw new Error('Incorrect password');
+  }
+
+  // return json web token
+  return {
+    token: jsonwebtoken.sign(
+      { user: { id: user.id, email: user.email } },
+      JWT_SECRET,
+      {
+        expiresIn: '1d'
+      }
+    )
+  };
 }
 
 module.exports = {
   Query: {
     users,
-    user
+    user,
+    myProfile
   },
   Mutation: {
-    newUser,
-    removeUser
+    removeUser,
+    signup,
+    login
   }
 };
